Sprint 4: Conclui US-13 com todos os testes passando, atualiza pom.xml com mockito-inline, atualiza documentação

+++ src/main/java/com/lbritosan/grokshop/dto/ProductDTO.java
@@ -1,15 +1,13 @@
 package com.lbritosan.grokshop.dto;
 
 import lombok.Data;
-
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class ProductDTO {
     private String name;
     private String description;
-    private BigDecimal price;
+    private Double price;
     private Integer stock;
     private List<Long> categoryIds;
 }
